organized lib folder
