+++ README.md
@@ -7,12 +7,11 @@
 [![License](https://img.shields.io/github/license/smaranjitghose/girlscript_app)](https://github.com/smaranjitghose/girlscript_app/blob/master/LICENSE)
 [![Gitter](https://badges.gitter.im/girlscript_app/community.svg)](https://gitter.im/girlscript_app/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge)
 
-
 A boilerplate for any technical community to pick up and use for their own without any hastle
-This app is a center to various events ranging from portraying the member’s achievements to announcing the upcoming opportunities. 
-Additionally, this app developed by the community, and for the community, will add elegance. Plus, being as an open source project, this will be a golden opportunity for beginners to learn and contribute to it just like a cherry on the cake. 
+This app is a center to various events ranging from portraying the member’s achievements to announcing the upcoming opportunities.
+Additionally, this app developed by the community, and for the community, will add elegance. Plus, being as an open source project, this will be a golden opportunity for beginners to learn and contribute to it just like a cherry on the cake.
 
-###  Mission
+### Mission
 
 - To empower people and specially under represented groups by educating them on technology and related necessary skills.
 - To create local opportunity ,growth and impact via tech education in every country around the world and bring diversity in tech domain.
@@ -25,13 +24,14 @@ Additionally, this app developed by the community, and for the community, will a
 - Introduce tech content that is helpful for everyone.
 
 ### Prerequisites to contribute
+
 - Ability to use Git and Github
 - Basic knowledge of Dart and Flutter
 - Basic knowledge of Firebase and willingness to learn more
 - UI/UX skills are appreciable but not mandatory
 
-
 ### Working and Screenshots
+
 This app opens up where you will see the option to sign up/log in to your account. After successfully logging in, you will have multiple tabs to go through. Each tab has a specific purpose. For example, the “Events” tab will list all the events taking place at certain places where you can register and attend those events.
 
 Following are some of the screenshots:
@@ -50,25 +50,50 @@ Following are some of the screenshots:
 
 </p>
 
+#### Folder structure
+
+`lib` : This is where overall logic of the application lies.
+
+`lib\models` : The Entire Data Models of the application stays here.
+
+`lib\services` : The Services of the app such as notification_service lies here.
+
+`lib\ui\screens` : The Whole Application Screens lies here.
+
+`lib\ui\custom` : The Custom Widgets you created can be placed here.
+
+#### Coding Guidelines
+
+1. Try to reduce the widget tree, in the long run it might help us a lot
+2. When it comes to package version, make sure you use the correct package.
+3. Try to seperate the UI from logic.
+4. Improve Documentation, this really helps to understand code.
+5. When it comes to naming convention, maintain similarity and stability through out the app.
+6. If your dart file has more than two widgets try to create sepearate file.
+7. When it comes to scale our application, try using `provider` or something similar to that.
+8. Know well about advanced concepts in flutter and try to use them properly.
+
 ### How to contribute:
 
 ![GSSOC'20](https://img.shields.io/badge/GSSOC-20-orange?style=for-the-badge)
 ![RGSOC'20](https://img.shields.io/badge/RGSOC-20-red?style=for-the-badge)
 
-
 **1.** Fork [this](https://github.com/smaranjitghose/girlscript_app) repository.
 
 **2.** Clone the forked repository.
+
 ```terminal
 git clone https://github.com/<your-github-username>/girlscript_app
 ```
 
 **3.** Navigate to the project directory.
+
 ```terminal
 cd girlscript_app
 ```
 
 **4.** Create a new branch.
+
 ```terminal
 git checkout -b <your_branch_name>
 ```
@@ -83,6 +108,7 @@ git checkout -b <your_branch_name>
 ```
 
 **7.** Push your local branch to the remote repository.
+
 ```terminal
 git push -u origin <your_branch_name>
 ```
@@ -91,35 +117,34 @@ git push -u origin <your_branch_name>
 
 **Congratulations!** Sit and relax, you've made your contribution to [GirlScript App](https://github.com/smaranjitghose/girlscript_app) project.
 
+### Resources to get started with:
 
-
-###  Resources to get started with:
-+  [Dart Programming](https://www.youtube.com/watch?v=Ej_Pcr4uC2Q)
-+  [Flutter Crash Course](https://www.youtube.com/watch?v=pTJJsmejUOQ)
-+  [Firebase and Flutter  App Build](https://www.youtube.com/playlist?list=PL4cUxeGkcC9j--TKIdkb3ISfRbJeJYQwC)
-+  [Reso Code's Youtube Channel for best practices](https://www.youtube.com/channel/UCSIvrn68cUk8CS8MbtBmBkA/featured)
-+  [Official Flutter Documentation](https://flutter.dev/docs)
-+  [Hyper Text Markup Language](https://www.w3schools.com/html/html_intro.asp)
-+ [Cascading Style Sheets](https://developer.mozilla.org/en-US/docs/Web/CSS/Reference)
-+ [JavaScript](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference)
-
+- [Dart Programming](https://www.youtube.com/watch?v=Ej_Pcr4uC2Q)
+- [Flutter Crash Course](https://www.youtube.com/watch?v=pTJJsmejUOQ)
+- [Firebase and Flutter App Build](https://www.youtube.com/playlist?list=PL4cUxeGkcC9j--TKIdkb3ISfRbJeJYQwC)
+- [Reso Code's Youtube Channel for best practices](https://www.youtube.com/channel/UCSIvrn68cUk8CS8MbtBmBkA/featured)
+- [Official Flutter Documentation](https://flutter.dev/docs)
+- [Hyper Text Markup Language](https://www.w3schools.com/html/html_intro.asp)
+- [Cascading Style Sheets](https://developer.mozilla.org/en-US/docs/Web/CSS/Reference)
+- [JavaScript](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference)
 
 ### Project Admin:
-* Smaranjit Ghose
 
- <a href="https://github.com/smaranjitghose"><img src="assets/Icons/GitHub_Mark.png" width=55></a>
+- Smaranjit Ghose
+
+<a href="https://github.com/smaranjitghose"><img src="assets/Icons/GitHub_Mark.png" width=55></a>
 ![](https://img.shields.io/twitter/url?label=Twitter&style=social&url=https%3A%2F%2Ftwitter.com%2Fsmaranjitghose)
 
 ### Mentors:
-* Sukkrit Sharma
- <a href="https://github.com/sukkritsharmaofficial"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
 
-* Saswat Nayak
-<a href="https://github.com/swat1998"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
+- Sukkrit Sharma
+  <a href="https://github.com/sukkritsharmaofficial"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
 
-* Anush Bhatia
-<a href="https://github.com/anushbhatia"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
+- Saswat Nayak
+  <a href="https://github.com/swat1998"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
 
-* Ousnik Polley
-<a href="https://github.com/ousnik"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
+- Anush Bhatia
+  <a href="https://github.com/anushbhatia"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
 
+- Ousnik Polley
+  <a href="https://github.com/ousnik"><img src="assets/Icons/GitHub_Mark.png" width=20></a>
