+++ community_app_boilerplate/lib/models/SocialIcons.dart
