+++ community_app_boilerplate/lib/ui/screens/EventRegistration.dart
